% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrasSolar.R
\name{BrasSolar}
\alias{BrasSolar}
\title{Bras solar}
\usage{
BrasSolar(el, R, nfac)
}
\arguments{
\item{el}{Solar elevation (deg from horizon).}

\item{R}{Distance from earth to sun in AU.}

\item{nfac}{Atmospheric turbidity parameter (2=clear, 4-5=smoggy").}
}
\value{
list
}
\description{
Bras method for atmospheric attenuation of solar radiation from a clear sky.
}
\details{
Called from Rshadexls script to get the Bras Solar equations calculations.
}
\seealso{
Other Earth: 
\code{\link{RyanStolzSolar}()},
\code{\link{SolarPosition}()},
\code{\link{calcHourAngleSunrise}()},
\code{\link{calcHourAngleSunset}()},
\code{\link{calcSolNoonUTC}()},
\code{\link{calcSunApparentLong}()},
\code{\link{calcSunRtAscension}()},
\code{\link{calcSunTrueLong}()},
\code{\link{solarazimuth}()},
\code{\link{solarelevation}()}
}
\concept{Earth}
