% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEccentricityEarthOrbit.R
\name{calcEccentricityEarthOrbit}
\alias{calcEccentricityEarthOrbit}
\title{Calculate Eccentricity of Earth Orbit}
\usage{
calcEccentricityEarthOrbit(t)
}
\arguments{
\item{t}{Number of Julian centuries since J2000.0.}
}
\value{
number (unit-less eccentricity)
}
\description{
Calculates the eccentricity of earth's orbit.
}
\seealso{
Other Solar system: 
\code{\link{calcEquationOfTime}()},
\code{\link{calcGeomMeanAnomalySun}()},
\code{\link{calcGeomMeanLongSun}()},
\code{\link{calcMeanObliquityOfEcliptic}()},
\code{\link{calcObliquityCorrection}()},
\code{\link{calcSunDeclination}()},
\code{\link{calcSunEqOfCenter}()},
\code{\link{calcSunRadVector}()},
\code{\link{calcSunTrueAnomaly}()}
}
\concept{Solar system}
