% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEquationOfTime.R
\name{calcEquationOfTime}
\alias{calcEquationOfTime}
\title{Calculate equation of time}
\usage{
calcEquationOfTime(t)
}
\arguments{
\item{t}{Number of Julian centuries since J2000.0}
}
\value{
Equation of time in minutes of time
}
\description{
Calculates the difference between true solar time and mean
}
\seealso{
Other Solar system: 
\code{\link{calcEccentricityEarthOrbit}()},
\code{\link{calcGeomMeanAnomalySun}()},
\code{\link{calcGeomMeanLongSun}()},
\code{\link{calcMeanObliquityOfEcliptic}()},
\code{\link{calcObliquityCorrection}()},
\code{\link{calcSunDeclination}()},
\code{\link{calcSunEqOfCenter}()},
\code{\link{calcSunRadVector}()},
\code{\link{calcSunTrueAnomaly}()}
}
\concept{Solar system}
