% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGeomMeanAnomalySun.R
\name{calcGeomMeanAnomalySun}
\alias{calcGeomMeanAnomalySun}
\title{Calculate Geometric Mean Anomaly of the Sun}
\usage{
calcGeomMeanAnomalySun(t)
}
\arguments{
\item{t}{Number of Julian centuries since J2000.0}
}
\value{
numeric (the Geometric Mean Anomaly of the Sun in degrees)
}
\description{
Calculates the Geometric Mean Anomaly of the Sun.
}
\seealso{
Other Solar system: 
\code{\link{calcEccentricityEarthOrbit}()},
\code{\link{calcEquationOfTime}()},
\code{\link{calcGeomMeanLongSun}()},
\code{\link{calcMeanObliquityOfEcliptic}()},
\code{\link{calcObliquityCorrection}()},
\code{\link{calcSunDeclination}()},
\code{\link{calcSunEqOfCenter}()},
\code{\link{calcSunRadVector}()},
\code{\link{calcSunTrueAnomaly}()}
}
\concept{Solar system}
