% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHourAngleSunrise.R
\name{calcHourAngleSunrise}
\alias{calcHourAngleSunrise}
\title{Calculate the Hour Angle of Sunrise}
\usage{
calcHourAngleSunrise(lat, SolarDec)
}
\arguments{
\item{lat}{Latitude of observer in degrees.}

\item{SolarDec}{Declination angle of sun in degrees}
}
\value{
numeric (hour angle of sunrise in radians)
}
\description{
Calculates the hour angle of the sun at sunrise for the latitude.
}
\seealso{
Other Earth: 
\code{\link{BrasSolar}()},
\code{\link{RyanStolzSolar}()},
\code{\link{SolarPosition}()},
\code{\link{calcHourAngleSunset}()},
\code{\link{calcSolNoonUTC}()},
\code{\link{calcSunApparentLong}()},
\code{\link{calcSunRtAscension}()},
\code{\link{calcSunTrueLong}()},
\code{\link{solarazimuth}()},
\code{\link{solarelevation}()}
}
\concept{Earth}
