% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHourAngleSunset.R
\name{calcHourAngleSunset}
\alias{calcHourAngleSunset}
\title{Calculate the Hour Angle of Sunset}
\usage{
calcHourAngleSunset(lat, SolarDec)
}
\arguments{
\item{lat}{Latitude of observer in degrees.}

\item{SolarDec}{Declination angle of sun in degrees}
}
\value{
numeric (hour angle of sunset in radians)
}
\description{
Calculates the hour angle of the sun at sunset for the latitude.
}
\seealso{
Other Earth: 
\code{\link{BrasSolar}()},
\code{\link{RyanStolzSolar}()},
\code{\link{SolarPosition}()},
\code{\link{calcHourAngleSunrise}()},
\code{\link{calcSolNoonUTC}()},
\code{\link{calcSunApparentLong}()},
\code{\link{calcSunRtAscension}()},
\code{\link{calcSunTrueLong}()},
\code{\link{solarazimuth}()},
\code{\link{solarelevation}()}
}
\concept{Earth}
