% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcJD.R
\name{calcJD}
\alias{calcJD}
\title{Julian day from calendar day}
\usage{
calcJD(date_year, date_month, date_day)
}
\arguments{
\item{date_year}{Numeric (4 digit year)}

\item{date_month}{Numeric (January = 1)}

\item{date_day}{Numeric (1 - 31)}
}
\value{
numeric (the Julian day corresponding to the date)
}
\description{
Julian day from calendar day.
}
\details{
Number is returned for start of day. Fractional days should be
}
\seealso{
Other Calendar: 
\code{\link{Julcvt}()},
\code{\link{calcJDFromJulianCent}()},
\code{\link{calcTimeJulianCent}()},
\code{\link{solarnoon}()}
}
\concept{Calendar}
