% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcJDFromJulianCent.R
\name{calcJDFromJulianCent}
\alias{calcJDFromJulianCent}
\title{Calculate Julian Day From Julian Century}
\usage{
calcJDFromJulianCent(t)
}
\arguments{
\item{t}{Number of Julian centuries since J2000.0.}
}
\value{
numeirc (the Julian Day corresponding to the t value)
}
\description{
Convert centuries since J2000.0 to Julian Day.
}
\seealso{
Other Calendar: 
\code{\link{Julcvt}()},
\code{\link{calcJD}()},
\code{\link{calcTimeJulianCent}()},
\code{\link{solarnoon}()}
}
\concept{Calendar}
