% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMeanObliquityOfEcliptic.R
\name{calcMeanObliquityOfEcliptic}
\alias{calcMeanObliquityOfEcliptic}
\title{Calculate mean obliquity of ecliptic}
\usage{
calcMeanObliquityOfEcliptic(t)
}
\arguments{
\item{t}{Number of Julian centuries since J2000.0.}
}
\value{
numeric (mean obliquity in degrees)
}
\description{
Calculate the mean obliquity of the ecliptic.
}
\seealso{
Other Solar system: 
\code{\link{calcEccentricityEarthOrbit}()},
\code{\link{calcEquationOfTime}()},
\code{\link{calcGeomMeanAnomalySun}()},
\code{\link{calcGeomMeanLongSun}()},
\code{\link{calcObliquityCorrection}()},
\code{\link{calcSunDeclination}()},
\code{\link{calcSunEqOfCenter}()},
\code{\link{calcSunRadVector}()},
\code{\link{calcSunTrueAnomaly}()}
}
\concept{Solar system}
