% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcObliquityCorrection.R
\name{calcObliquityCorrection}
\alias{calcObliquityCorrection}
\title{Calculate Obliquity Correction}
\usage{
calcObliquityCorrection(t)
}
\arguments{
\item{t}{Number of Julian centuries since J2000.0.}
}
\value{
numeric (corrected obliquity in degrees)
}
\description{
Calculate the corrected obliquity of the ecliptic.
}
\seealso{
Other Solar system: 
\code{\link{calcEccentricityEarthOrbit}()},
\code{\link{calcEquationOfTime}()},
\code{\link{calcGeomMeanAnomalySun}()},
\code{\link{calcGeomMeanLongSun}()},
\code{\link{calcMeanObliquityOfEcliptic}()},
\code{\link{calcSunDeclination}()},
\code{\link{calcSunEqOfCenter}()},
\code{\link{calcSunRadVector}()},
\code{\link{calcSunTrueAnomaly}()}
}
\concept{Solar system}
