% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSolNoonUTC.R
\name{calcSolNoonUTC}
\alias{calcSolNoonUTC}
\title{Calculate Solar Noon UTC}
\usage{
calcSolNoonUTC(t, longitude)
}
\arguments{
\item{t}{Number of Julian centuries since J2000.0}

\item{longitude}{Longitude of observer in degrees.}
}
\value{
numeric (time in minutes from zero Z)
}
\description{
Calculate the Universal Coordinated Time (UTC) of solar
noon for the given day at the given location on earth.
}
\seealso{
Other Earth: 
\code{\link{BrasSolar}()},
\code{\link{RyanStolzSolar}()},
\code{\link{SolarPosition}()},
\code{\link{calcHourAngleSunrise}()},
\code{\link{calcHourAngleSunset}()},
\code{\link{calcSunApparentLong}()},
\code{\link{calcSunRtAscension}()},
\code{\link{calcSunTrueLong}()},
\code{\link{solarazimuth}()},
\code{\link{solarelevation}()}
}
\concept{Earth}
