% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSunApparentLong.R
\name{calcSunApparentLong}
\alias{calcSunApparentLong}
\title{Calculate Sun Apparent Longitude}
\usage{
calcSunApparentLong(t)
}
\arguments{
\item{t}{Number of Julian centuries since J2000.0.}
}
\value{
numeric (sun's apparent longitude in degrees)
}
\description{
Calculate the apparent longitude of the sun.
}
\details{
Not used by sunrise, solar noon, sunset.
}
\seealso{
Other Earth: 
\code{\link{BrasSolar}()},
\code{\link{RyanStolzSolar}()},
\code{\link{SolarPosition}()},
\code{\link{calcHourAngleSunrise}()},
\code{\link{calcHourAngleSunset}()},
\code{\link{calcSolNoonUTC}()},
\code{\link{calcSunRtAscension}()},
\code{\link{calcSunTrueLong}()},
\code{\link{solarazimuth}()},
\code{\link{solarelevation}()}
}
\concept{Earth}
