% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSunDeclination.R
\name{calcSunDeclination}
\alias{calcSunDeclination}
\title{Calculate Sun Declination}
\usage{
calcSunDeclination(t)
}
\arguments{
\item{t}{Number of Julian centuries since J2000.0.}
}
\value{
numeric (sun's declination in degrees)
}
\description{
Calculate the declination of the sun.
}
\seealso{
Other Solar system: 
\code{\link{calcEccentricityEarthOrbit}()},
\code{\link{calcEquationOfTime}()},
\code{\link{calcGeomMeanAnomalySun}()},
\code{\link{calcGeomMeanLongSun}()},
\code{\link{calcMeanObliquityOfEcliptic}()},
\code{\link{calcObliquityCorrection}()},
\code{\link{calcSunEqOfCenter}()},
\code{\link{calcSunRadVector}()},
\code{\link{calcSunTrueAnomaly}()}
}
\concept{Solar system}
