% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSunEqOfCenter.R
\name{calcSunEqOfCenter}
\alias{calcSunEqOfCenter}
\title{Calculate Sun Equation of Of Center for the Sun}
\usage{
calcSunEqOfCenter(t)
}
\arguments{
\item{t}{Number of Julian centuries since J2000.0.}
}
\value{
numeric (degrees)
}
\description{
Calculate the equation of center for the Sun.
}
\seealso{
Other Solar system: 
\code{\link{calcEccentricityEarthOrbit}()},
\code{\link{calcEquationOfTime}()},
\code{\link{calcGeomMeanAnomalySun}()},
\code{\link{calcGeomMeanLongSun}()},
\code{\link{calcMeanObliquityOfEcliptic}()},
\code{\link{calcObliquityCorrection}()},
\code{\link{calcSunDeclination}()},
\code{\link{calcSunRadVector}()},
\code{\link{calcSunTrueAnomaly}()}
}
\concept{Solar system}
