% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSunRadVector.R
\name{calcSunRadVector}
\alias{calcSunRadVector}
\title{Calculate Sun Radius Vector}
\usage{
calcSunRadVector(t)
}
\arguments{
\item{t}{Number of Julian centuries since J2000.0.}
}
\value{
numeric (Sun radius vector in AUs)
}
\description{
Calculate the distance to the sun in AU.
}
\details{
(Not used by sunrise, solar noon, sunset.)
}
\seealso{
Other Solar system: 
\code{\link{calcEccentricityEarthOrbit}()},
\code{\link{calcEquationOfTime}()},
\code{\link{calcGeomMeanAnomalySun}()},
\code{\link{calcGeomMeanLongSun}()},
\code{\link{calcMeanObliquityOfEcliptic}()},
\code{\link{calcObliquityCorrection}()},
\code{\link{calcSunDeclination}()},
\code{\link{calcSunEqOfCenter}()},
\code{\link{calcSunTrueAnomaly}()}
}
\concept{Solar system}
