% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSunriseUTC.R
\name{calcSunriseUTC}
\alias{calcSunriseUTC}
\title{Calculate the Sunrise UTC}
\usage{
calcSunriseUTC(jd, Latitude, longitude)
}
\arguments{
\item{jd}{Julian day.}

\item{Latitude}{Latitude of observer in degrees.}

\item{longitude}{Longitude of observer in degrees.}
}
\value{
numeric (time in minutes from zero Z)
}
\description{
Calculates the Universal Coordinated Time (UTC) of sunrise
for the given day at the given location on earth
}
\seealso{
Other Earth, Calendar: 
\code{\link{calcSunsetUTC}()},
\code{\link{sunrise}()},
\code{\link{sunset}()}
}
\concept{Earth, Calendar}
