% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSunRtAscension.R
\name{calcSunRtAscension}
\alias{calcSunRtAscension}
\title{Calculate Sun Right Ascension}
\usage{
calcSunRtAscension(t)
}
\arguments{
\item{t}{Number of Julian centuries since J2000.0}
}
\value{
numeric (Sun's right ascension in degrees)
}
\description{
Calculate the right ascension of the sun.
}
\details{
(Not used by sunrise, solar noon, sunset.)
}
\seealso{
Other Earth: 
\code{\link{BrasSolar}()},
\code{\link{RyanStolzSolar}()},
\code{\link{SolarPosition}()},
\code{\link{calcHourAngleSunrise}()},
\code{\link{calcHourAngleSunset}()},
\code{\link{calcSolNoonUTC}()},
\code{\link{calcSunApparentLong}()},
\code{\link{calcSunTrueLong}()},
\code{\link{solarazimuth}()},
\code{\link{solarelevation}()}
}
\concept{Earth}
