% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSunsetUTC.R
\name{calcSunsetUTC}
\alias{calcSunsetUTC}
\title{Calculate Sunset UTC}
\usage{
calcSunsetUTC(jd, Latitude, longitude)
}
\arguments{
\item{jd}{Julian day.}

\item{Latitude}{Latitude of observer in degrees.}

\item{longitude}{Longitude of observer in degrees.}
}
\value{
numeric (time in minutes from zero Z)
}
\description{
Calculate the Universal Coordinated Time (UTC) of sunset
for the given day at the given location on Earth.
}
\seealso{
Other Earth, Calendar: 
\code{\link{calcSunriseUTC}()},
\code{\link{sunrise}()},
\code{\link{sunset}()}
}
\concept{Earth, Calendar}
