% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSunTrueAnomaly.R
\name{calcSunTrueAnomaly}
\alias{calcSunTrueAnomaly}
\title{Calculate Sun True Anomaly}
\usage{
calcSunTrueAnomaly(t)
}
\arguments{
\item{t}{Number of Julian centuries since J2000.0.}
}
\value{
numeric (Sun's true anomaly in degrees)
}
\description{
Calculate the true anomaly of the Sun.
}
\seealso{
Other Solar system: 
\code{\link{calcEccentricityEarthOrbit}()},
\code{\link{calcEquationOfTime}()},
\code{\link{calcGeomMeanAnomalySun}()},
\code{\link{calcGeomMeanLongSun}()},
\code{\link{calcMeanObliquityOfEcliptic}()},
\code{\link{calcObliquityCorrection}()},
\code{\link{calcSunDeclination}()},
\code{\link{calcSunEqOfCenter}()},
\code{\link{calcSunRadVector}()}
}
\concept{Solar system}
