% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSunTrueLong.R
\name{calcSunTrueLong}
\alias{calcSunTrueLong}
\title{Calculate Sun True Longitude}
\usage{
calcSunTrueLong(t)
}
\arguments{
\item{t}{Number of Julian centuries since J2000.0.}
}
\value{
numeric (Sun's true longitude in degrees)
}
\description{
Calculate the true longitude of the Sun.
}
\seealso{
Other Earth: 
\code{\link{BrasSolar}()},
\code{\link{RyanStolzSolar}()},
\code{\link{SolarPosition}()},
\code{\link{calcHourAngleSunrise}()},
\code{\link{calcHourAngleSunset}()},
\code{\link{calcSolNoonUTC}()},
\code{\link{calcSunApparentLong}()},
\code{\link{calcSunRtAscension}()},
\code{\link{solarazimuth}()},
\code{\link{solarelevation}()}
}
\concept{Earth}
