% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTimeJulianCent.R
\name{calcTimeJulianCent}
\alias{calcTimeJulianCent}
\title{Calculate Time Julian Centuries}
\usage{
calcTimeJulianCent(jd)
}
\arguments{
\item{jd}{Julian Day to convert}
}
\value{
numeric (the T value corresponding to the Julian Day)
}
\description{
Convert Julian Day to centuries since J2000.0.
}
\seealso{
Other Calendar: 
\code{\link{Julcvt}()},
\code{\link{calcJDFromJulianCent}()},
\code{\link{calcJD}()},
\code{\link{solarnoon}()}
}
\concept{Calendar}
