% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degTorad.R
\name{degToRad}
\alias{degToRad}
\title{Degrees to Radians}
\usage{
degToRad(x)
}
\arguments{
\item{x}{Angle in degrees}
}
\value{
numeric (an angle in radians)
}
\description{
Function for converting from degrees to radians.
}
\seealso{
Other Data wrangling: 
\code{\link{dfcolumns2robjects}()},
\code{\link{transverse_matrix_fnx}()}
}
\concept{Data wrangling}
