% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfcolumns2robjects.R
\name{dfcolumns2robjects}
\alias{dfcolumns2robjects}
\title{Data frame columns to R-objects}
\usage{
dfcolumns2robjects(data_frame_sub = site_dat, ...)
}
\arguments{
\item{data_frame_sub}{Data frame}

\item{...}{Not used}
}
\value{
environment
}
\description{
Function to pull site data from data frame and load them as R objects.
}
\seealso{
Other Data wrangling: 
\code{\link{degToRad}()},
\code{\link{transverse_matrix_fnx}()}
}
\concept{Data wrangling}
