% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfSmallSiteData}
\alias{dfSmallSiteData}
\title{Small site data}
\format{
A data frame with 20 rows and 172 columns.
}
\source{
EPA.
}
\usage{
dfSmallSiteData
}
\description{
Site data provided with the initial test runs.
}
\keyword{datasets}
