% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs7Solar.R
\name{hs7Solar}
\alias{hs7Solar}
\title{HS 7 solar}
\usage{
hs7Solar(theTime, Hour_DST, theElevation, SolarAltitude)
}
\arguments{
\item{theTime}{Time.}

\item{Hour_DST}{Hour.}

\item{theElevation}{Elevation.}

\item{SolarAltitude}{Solar altitude.}
}
\value{
list
}
\description{
HeatSource version 7 solar flux calculation before shade from topo and veg.
}
