% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs-fixed-star.R
\name{inputs_fixed_star}
\alias{inputs_fixed_star}
\title{Inputs for "fixed star"}
\usage{
inputs_fixed_star(
  site_data_frame = site_data,
  site_id_chr = "pid",
  lat_chr = "lat",
  lon_chr = "lon",
  startdate = "2001-08-01",
  ndays = 31,
  timezone = -8,
  dlstime = 1,
  VWidthGlobal = 30.48,
  JsntMethod = "Bras",
  cloudfrac_global = 0,
  nfac = 2,
  atc = 0.8,
  vegzones = "9",
  star_veg = TRUE,
  ripgelev = FALSE
)
}
\arguments{
\item{site_data_frame}{Holds all the site details necessary for the inputs_fixed()
fnx to process and develop the necessary objects for shade
     [this data frame has the following fields:
          site_id: numerical vlaue to identify each site
          lat: latitude of site in decimal degrees
          lon: longitude of site in decimal degrees
          zone_elev[1:X]: reach-specific ground elevation for each
               riparian vegetation buffer zone at a site
          zone_wid[1:X]: reach-specific riparian zone widths for each
               riparian vegetation buffer zone at a site}

\item{site_id_chr}{site_id_chr}

\item{lat_chr}{Latitude.}

\item{lon_chr}{Longitude.}

\item{startdate}{Start date for integrating shade as a text string "YYYY-MM-DD"}

\item{ndays}{Number of days to integrate/simulate shade starting on 'startdate'
For example, to simulate shade only on the start date enter 1.
To simulate shade for the entire month of August 2012 enter 'startdate'
"2012-08-01" and 'ndays' = 31.
nday of 1 corresponds to simulating just the 'startdate'}

\item{timezone}{Time zone is the difference between UTC and the local standard
time in hours. A map of world time zones may be found at:
http://aa.usno.navy.mil/graphics/TimeZoneMap2001.pdf
PST is -8 for the Pacific Northwest USA.}

\item{dlstime}{This is a daylight savings time indicator variable where you Select '1'
if you want times calculated and reported as local daylight savings
time or "0" to report times in local standard time.}

\item{JsntMethod}{Select the solar model to generate input radiation.
options selected from: 0="Bras" 3="ODEQ" 4="Ryan-Stolz".}

\item{cloudfrac_global}{Global value for cloud fraction of total incoming energy
that will be applied to all times. If continuous hourly values are
entered in the site_data_frame, then they override this global value.
Value must be between 0 and 1.}

\item{nfac}{Atmospheric turbidity factor for calculation of solar radiation that
is used if the Bras option for the solar radiation model is selected.
[2 = clear, 5 = smog].}

\item{atc}{Atmospheric transmission factor for calculation of solar radiation
used if the Ryan-Stolzenbach option for 'JsntMethod' is selected.}

\item{vegzones}{vegzones}

\item{star_veg}{star_veg}

\item{ripgelev}{ripgelev}

\item{VWidthGlobalEach}{of the 9 riparian vegetation zones on each bank are
assumed to have the same width (meters) as specified here. set to NA
if specific zone widths are desired for the ...}
}
\value{
Data frame
}
\description{
Function to load all fixed site input variables/values.
}
\concept{Data ingestion}
