% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Julcvt.R
\name{Julcvt}
\alias{Julcvt}
\title{Julcvt}
\usage{
Julcvt(mo, da, yr)
}
\arguments{
\item{mo}{Month.}

\item{da}{Day.}

\item{yr}{Year.}
}
\description{
Calculates the day of the year.
}
\seealso{
Other Calendar: 
\code{\link{calcJDFromJulianCent}()},
\code{\link{calcJD}()},
\code{\link{calcTimeJulianCent}()},
\code{\link{solarnoon}()}
}
\concept{Calendar}
