% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chen-and-Parchunk-star-method.R
\name{method_chen98_parchunk_star}
\alias{method_chen98_parchunk_star}
\title{Chen and Parchunk star method}
\usage{
method_chen98_parchunk_star(
  fixed_site_df = fixed_site_inputs,
  time_interval = 60,
  extOpt = 0,
  cloudfrac_flag = "global",
  shddenadd = 0.2,
  ovhvegadd = 1,
  smpl_dir = "three",
  star_veg = FALSE,
  vegzones = 1,
  par_processes = 2,
  profilingQ = TRUE,
  ...
)
}
\arguments{
\item{fixed_site_df}{Sample data is given as a default value.}

\item{time_interval}{Integration interval in minutes (60 = hourly).}

\item{extOpt}{Use Beer's Law.}

\item{cloudfrac_flag}{Default uses global cloud fraction.}

\item{shddenadd}{Chen et al. 1998 model 20% constant pg311.}

\item{ovhvegadd}{ovhvegadd}

\item{smpl_dir}{smpl_dir}

\item{star_veg}{star_veg}

\item{vegzones}{vegzones}

\item{par_processes}{Number of parallel processes to run.}

\item{profilingQ}{Should profiling information be printed out or not.}
}
\value{
data frame
}
\description{
Chen and Parchunk star method
}
\details{
Chen et al 1998 SHADE model methods derived from the Washington Ecology
Department 'shade.xls' model. Code translated from VBA Calc_FLUX sub model.
optional calculation using Chen or ODEQ method (0=Chen method, 1=ODEQ method)

The Chen method is recommended for use with the QUAL2K model. The Chen
method is documented in the PhD dissertation by Y.D. Chen and publications
in the Journal of Environmental Engineering:

Chen, Y.D. (1996). Hydrologic and water quality modeling for aquatic
ecosystem protection and restoration in forest watersheds: a case study
of stream temperature in the Upper Grande Ronde River, Oregon.
PhD dissertation. University of Georgia. Athens, GA.

Chen, Y.D.,  Carsel, R.F., McCutcheon, S.C., and Nutter, W.L. (1998).
Stream temperature simulation of forested riparian areas: I.
watershed-scale model development. Journal of Environmental Engineering.
April 1998. pp 304-315.

Chen, Y.D.,  Carsel, R.F., McCutcheon, S.C., and Nutter, W.L. (1998).
Stream temperature simulation of forested riparian areas: II. model
application. Journal of Environmental Engineering. April 1998. pp 316-328.
}
\concept{Method}
