% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radTodeg.R
\name{radToDeg}
\alias{radToDeg}
\title{Radians to degrees}
\usage{
radToDeg(x)
}
\arguments{
\item{x}{Radians.}
}
\value{
numeric (degrees)
}
\description{
Radians to degrees
}
\seealso{
Other Data wranling: 
\code{\link{veg_elev_matrix_fnx}()}
}
\concept{Data wranling}
