% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyanStolzSolar.R
\name{RyanStolzSolar}
\alias{RyanStolzSolar}
\title{Ryan-Stolzenbach method}
\usage{
RyanStolzSolar(el, R, atc, z, Iclear, I0)
}
\arguments{
\item{el}{Solar elevation deg from horizon.}

\item{R}{Distance from earth to sun in AU.}

\item{atc}{Atmospheric transmission coefficient (0.70-0.91, default 0.8)}

\item{z}{Elevation, metres -- required if imthd=2.}

\item{Iclear}{Clear-sky solar radiation, W m-2.}

\item{I0}{Clear-sky solar radiation at top of atmosphere, W m-2.}
}
\value{
list
}
\description{
Ryan-Stolzenbach method for atmospheric attenuation of solar radiation from a clear sky.
}
\seealso{
Other Earth: 
\code{\link{BrasSolar}()},
\code{\link{SolarPosition}()},
\code{\link{calcHourAngleSunrise}()},
\code{\link{calcHourAngleSunset}()},
\code{\link{calcSolNoonUTC}()},
\code{\link{calcSunApparentLong}()},
\code{\link{calcSunRtAscension}()},
\code{\link{calcSunTrueLong}()},
\code{\link{solarazimuth}()},
\code{\link{solarelevation}()}
}
\concept{Earth}
