% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKOP_fnx.R
\name{SKOP_fnx}
\alias{SKOP_fnx}
\title{Sky Opening function}
\usage{
SKOP_fnx(veg_elev_matrix, transverse_matrix)
}
\arguments{
\item{veg_elev_matrix}{Matrix rows = 2 1-left, 2-right; cols = # vegzones;
values = ground elevation for each zone.}

\item{transverse_matrix}{Matrix rows = 2 1-left, 2-right; cols = # vegzones;
values = distance from center of stream to the edge of each zone.}
}
\value{
numeric (SKOP value)
}
\description{
Sky Opening function
}
