% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarelevation.R
\name{solarelevation}
\alias{solarelevation}
\title{Solar elevation}
\usage{
solarelevation(
  lat,
  lon,
  year,
  month,
  day,
  hours,
  minutes,
  seconds,
  timezone,
  dlstime
)
}
\arguments{
\item{lat}{Latitude.}

\item{lon}{Longitude.}

\item{year}{Year.}

\item{month}{Month.}

\item{day}{Day.}

\item{hours}{Hours.}

\item{minutes}{Minutes.}

\item{seconds}{Seconds.}

\item{timezone}{Time zone.}

\item{dlstime}{Daylight saving time}
}
\value{
numeric (solar azimuth in degrees from north)
}
\description{
Calculate solar azimuth (deg from north) for the entered
date, time and location. Returns -999999 if darker than twilight.
}
\details{
The functions solarelevation and solarazimuth are identical
and could be converted to a VBA subroutine that would return
both values.
}
\seealso{
Other Earth: 
\code{\link{BrasSolar}()},
\code{\link{RyanStolzSolar}()},
\code{\link{SolarPosition}()},
\code{\link{calcHourAngleSunrise}()},
\code{\link{calcHourAngleSunset}()},
\code{\link{calcSolNoonUTC}()},
\code{\link{calcSunApparentLong}()},
\code{\link{calcSunRtAscension}()},
\code{\link{calcSunTrueLong}()},
\code{\link{solarazimuth}()}
}
\concept{Earth}
