% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarnoon.R
\name{solarnoon}
\alias{solarnoon}
\title{Solar noon}
\usage{
solarnoon(lat, lon, year, month, day, timezone, dlstime)
}
\arguments{
\item{lat}{Latitude.}

\item{lon}{Longitude.}

\item{year}{Year.}

\item{month}{Month.}

\item{day}{Day.}

\item{timezone}{Time zone.}

\item{dlstime}{Daylight saving time,}
}
\value{
numeric (time of solar noon in local time days)
}
\description{
Calculate the Universal Coordinated Time (UTC) of solar
noon for the given day at the given location on Earth.
}
\details{
Main Function called by spreadsheet.
}
\seealso{
Other Calendar: 
\code{\link{Julcvt}()},
\code{\link{calcJDFromJulianCent}()},
\code{\link{calcJD}()},
\code{\link{calcTimeJulianCent}()}
}
\concept{Calendar}
