% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolarPosition.R
\name{SolarPosition}
\alias{SolarPosition}
\title{Solar Position}
\usage{
SolarPosition(
  lat,
  lon,
  date_year,
  date_month,
  date_day,
  thours,
  tminutes,
  tseconds,
  timezone,
  dlstime
)
}
\arguments{
\item{lat}{Latitude.}

\item{lon}{Longitude.}

\item{date_year}{Year.}

\item{date_month}{Month.}

\item{date_day}{Day.}

\item{thours}{Hours.}

\item{tminutes}{Minutes.}

\item{tseconds}{Seconds.}

\item{timezone}{Time zone.}

\item{dlstime}{Daylight saving time.}
}
\value{
numeric (solar azimuth in degrees from north)
}
\description{
Solar Position
}
\details{
solarelevation and solarazimuth functions are identical
and could converted to a VBA subroutine that would return
both values.
}
\seealso{
Other Earth: 
\code{\link{BrasSolar}()},
\code{\link{RyanStolzSolar}()},
\code{\link{calcHourAngleSunrise}()},
\code{\link{calcHourAngleSunset}()},
\code{\link{calcSolNoonUTC}()},
\code{\link{calcSunApparentLong}()},
\code{\link{calcSunRtAscension}()},
\code{\link{calcSunTrueLong}()},
\code{\link{solarazimuth}()},
\code{\link{solarelevation}()}
}
\concept{Earth}
