% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transverse_matrix_fnx.R
\name{transverse_matrix_fnx}
\alias{transverse_matrix_fnx}
\title{Transverse distance to vegetation zone calculations}
\usage{
transverse_matrix_fnx(lbk_vwidth, rbk_vwidth, vegzones)
}
\arguments{
\item{lbk_vwidth}{A vector holding vegetation zone widths for left bank zones.}

\item{rbk_vwidth}{A vector holding vegetation zone widths for right bank zones.}
}
\value{
matrix holding left/right bank transverse distances to veg zone edges
}
\description{
Transverse distance to vegetation zone calculations
}
\seealso{
Other Data wrangling: 
\code{\link{degToRad}()},
\code{\link{dfcolumns2robjects}()}
}
\concept{Data wrangling}
