% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_elev_matrix_fnx.R
\name{veg_elev_matrix_fnx}
\alias{veg_elev_matrix_fnx}
\title{Elevation of vegetation in each zone}
\usage{
veg_elev_matrix_fnx(
  lbk_grdelev,
  rbk_grdelev,
  lbk_ripHght,
  rbk_ripHght,
  vegzones,
  elevation
)
}
\arguments{
\item{lbk_grdelev}{A vector holding vegetation zone ground elevation for left bank zones.}

\item{rbk_grdelev}{A vector holding vegetation zone ground elevation for right bank zones.}

\item{lbk_ripHght}{A vector holding vegetation zone vegetation height for left bank zones.}

\item{rbk_ripHght}{A vector holding vegetation zone vegetation height for left bank zones.}

\item{vegzones}{A numerical value holding the number of vegetation zones for the site.}

\item{elevation}{A numerical value holding the elevation of the site.}
}
\value{
matrix holding left/right bank transverse distances to veg zone edges
}
\description{
Elevation of vegetation in each zone
}
\seealso{
Other Data wranling: 
\code{\link{radToDeg}()}
}
\concept{Data wranling}
